AV2-249: Model for export queues to CSV -refactoring
